# lib/api-zod/src/generated/types/gateView.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GateViewGateType } from "./gateViewGateType";
import type { GateViewState } from "./gateViewState";

export interface GateView {
  code: string;
  name: string;
  gateType: GateViewGateType;
  locks: string;
  opensWhen: string;
  openedBy: string;
  state: GateViewState;
}