# lib/api-zod/src/generated/types/meResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentRecord } from "./documentRecord";
import type { GateView } from "./gateView";
import type { ProfileView } from "./profileView";
import type { StatusStep } from "./statusStep";
import type { SubmissionSummary } from "./submissionSummary";

export interface MeResponse {
  profile: ProfileView;
  gates: GateView[];
  statuses: StatusStep[];
  documents: DocumentRecord[];
  submissions: SubmissionSummary[];
  canApply: boolean;
  applicationSubmitted: boolean;
  showIntroductionForm: boolean;
}