# lib/api-zod/src/generated/types/submissionSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubmissionField } from "./submissionField";
import type { SubmissionSummaryFormType } from "./submissionSummaryFormType";

export interface SubmissionSummary {
  id: number;
  formType: SubmissionSummaryFormType;
  title: string;
  submittedAt: Date;
  details: SubmissionField[];
}